+++ gym_hil/__init__.py
@@ -90,13 +90,32 @@ register(
 )
 
 register(
+    id="gym_hil/PandaGuessTheOrder-v0",
+    entry_point="gym_hil.wrappers.factory:make_env",
+    max_episode_steps=100,
+    kwargs={
+        "env_id": "gym_hil/PandaPickCubeBase-v0",  # Use the base environment
+    },
+)
+
+register(
+    id="gym_hil/PandaGuessTheOrderGamepad-v0",
+    entry_point="gym_hil.wrappers.factory:make_env",
+    max_episode_steps=100,
+    kwargs={
+        "env_id": "gym_hil/PandaGuessTheOrderBase-v0",  # Use the base environment
+        "use_viewer": True,
+        "use_gamepad": True,
+    },
+)
+
+register(
     id="gym_hil/PandaGuessTheOrderKeyboard-v0",
     entry_point="gym_hil.wrappers.factory:make_env",
     max_episode_steps=100,
     kwargs={
         "env_id": "gym_hil/PandaGuessTheOrderBase-v0",
         "use_viewer": True,
-        "gripper_penalty": -0.05,
-        "use_gamepad": False
+        "gripper_penalty": -0.05
     }
 )
