+++ gym_hil/assets/scene2.xml
@@ -48,23 +48,23 @@
       </body>
 
       <!-- Targets-->
-      <body name="target1" pos="-0.7 0.5 .02">
+      <body name="target1" pos="0.5 -0.7  .02">
         <freejoint name="target1"/>
         <geom name="target1" type="box" size=".05 .05 .005" mass="0.1"  rgba="1.0 .0 .0 1"/>
       </body>
-      <body name="target2" pos="-0.35 0.5 .02">
+      <body name="target2" pos=" 0.5 -0.35 .02">
         <freejoint name="target2"/>
         <geom name="target2" type="box" size=".05 .05 .005" mass="0.1"  rgba=".0 .0 1.0 1"/>
       </body>
-      <body name="target3" pos="0 0.5 .02">
+      <body name="target3" pos="0.5 0 .02">
         <freejoint name="target3"/>
         <geom name="target3" type="box" size=".05 .05 .005"  mass="0.1"  rgba=".1 .3 .1 1"/>
       </body>
-      <body name="target4" pos="0.35 0.5 .02">
+      <body name="target4" pos="0.5 0.35 .02">
         <freejoint name="target4"/>
         <geom name="target4" type="box" size=".05 .05 .005" mass="0.1"  rgba=".2 .66 .32 1"/>
       </body>
-      <body name="target5" pos="0.7 0.5 .02">
+      <body name="target5" pos="0.5 0.7 .02">
         <freejoint name="target5"/>
         <geom name="target5" type="box" size=".05 .05 .005" mass="0.1"  rgba=".6 .3 .6 1"/>
       </body>
